requirements.txt
djangorestframework==3.12.4
django-cors-headers==3.8.0
beautifulsoup4
html5lib
redis
Celery